Cube fit and coordinate conversion separated and made into their own files
